requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-ckeditor==6.0.0
django-dotenv==1.4.2
django-js-asset==1.2.2
django-mptt==0.11.0
flower==0.9.5
humanize==3.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
pdfkit==0.6.1
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
python-http-client==3.3.1
pytz==2020.4
redis==3.5.3
sendgrid==6.4.8
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
wkhtmltopdf==0.2
